+++ Teste3/Atividade8Lista3/main.c
@@ -3,14 +3,14 @@
 
 int main()
 {
-     int x = 1;
+     float x = 1;
     float soma = 0;
 
   while (x <= 10)
     {
         soma += (1 / x);
          x++;
-        
+
     }
 
      printf("O resultado e': %.2f", soma);
